requirements.txt
psycopg2-binary
python-decouple
fastapi
uvicorn
pandas
psycopg2
sqlalchemy
openpyxl
python-dotenv
